@@ -1,6 +1,6 @@
 {
   "name": "docu",
-  "version": "0.0.0",
+  "version": "0.0.1",
   "private": true,
   "scripts": {
     "docusaurus": "docusaurus",
@@ -14,18 +14,24 @@
     "write-heading-ids": "docusaurus write-heading-ids"
   },
   "dependencies": {
-    "@cmfcmf/docusaurus-search-local": "1.1.0",
-    "@docusaurus/core": "^2.4",
-    "@docusaurus/plugin-client-redirects": "^2.4.1",
-    "@docusaurus/plugin-google-tag-manager": "^2.3.1",
-    "@docusaurus/preset-classic": "2.4",
-    "@mdx-js/react": "^1.6.21",
+    "@docusaurus/core": "3.5.2",
+    "@docusaurus/plugin-client-redirects": "3.5.2",
+    "@docusaurus/plugin-google-tag-manager": "3.5.2",
+    "@docusaurus/preset-classic": "3.5.2",
+    "@mdx-js/react": "^3.0.1",
     "clsx": "^1.1.1",
     "docusaurus-plugin-hubspot": "^1.0.0",
     "docusaurus-plugin-segment": "^1.0.3",
-    "prism-react-renderer": "^1.2.1",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1"
+    "prism-react-renderer": "^2.1.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
+  },
+  "devDependencies": {
+    "@docusaurus/module-type-aliases": "3.5.2",
+    "@docusaurus/types": "3.5.2"
+  },
+  "engines": {
+    "node": ">=18.0"
   },
   "resolutions": {
     "cheerio": "1.0.0-rc.12"
